pyproject.toml
[project]
name = "investment-guide-project-ai"
version = "0.1.0"
description = ""
authors = [
    {name = "abhay-byte",email = "abhay02delhi@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit (>=1.44.1,<2.0.0)",
    "numpy (>=2.2.4,<3.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "seaborn (>=0.13.2,<0.14.0)",
    "plotly (>=6.0.1,<7.0.0)",
    "google-genai (>=1.10.0,<2.0.0)",
    "selenium (>=4.31.0,<5.0.0)",
    "webdriver-manager (>=4.0.2,<5.0.0)",
    "bs4 (>=0.0.2,<0.0.3)",
    "python-dotenv",
    "feedparser",
    "yfinance",
    "fpdf",
    "PyPDF2",
    "streamlit-chat"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
